pyproject.toml
[project]
name = "meetuper-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "uvloop (>=0.21.0,<0.22.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
